requirements.txt
playwright>=1.50.0
pyyaml>=6.0
aiosqlite>=0.20.0
aiosmtplib>=3.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
